check.py
import os
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()
gemini_api_key = os.getenv("GEMINI_API_KEY")

client = OpenAI(
    api_key=gemini_api_key,
    base_url="https://generativelanguage.googleapis.com/v1beta/openai/"
)

# List models synchronously
response = client.models.list()
for m in response.data:
    print(m.id)
