pyproject.toml
[project]
name = "book-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cohere>=5.20.0",
    "fastapi[standard]>=0.124.4",
    "openai-agents>=0.6.2",
    "qdrant-client>=1.16.1",
    "trafilatura>=2.0.0",
]
